+++ plugins/Clist_modern/src/modern_clcutils.cpp
@@ -724,7 +724,7 @@ int ExpandMetaContact(HWND hwnd, ClcContact *contact, ClcData *dat, BOOL bExpand
 int cliFindRowByText(HWND hwnd, ClcData *dat, const TCHAR *text, int prefixOk)
 {
 	ClcGroup *group = &dat->list;
-	int testlen = mir_tstrlen(text);
+	int testlen = (int)mir_tstrlen(text);
 	ClcContact *contact = NULL;
 	int SubCount = 0;
